+++ Editor/Content/Asset.cs
@@ -35,6 +35,7 @@ namespace Editor.Content
         public abstract IEnumerable<string> Save(string file);
         public abstract void Import(string file);
         public abstract void Load(string file);
+        public abstract byte[] PackForEngine();
 
         public static AssetInfo GetAssetInfo(string file)
         {