+++ Editor/Utilities/MathUtilities.cs
@@ -1,4 +1,6 @@
-﻿namespace Editor.Utilities
+﻿using System.Diagnostics;
+
+namespace Editor.Utilities
 {
     public static class MathUtilities
     {
@@ -11,6 +13,28 @@
             if (!value.HasValue || !other.HasValue) return false;
 
             return IsEqual(value.Value, other.Value);
+        }
+
+        public static long AlignSizeUp(long size, long alignment)
+        {
+            Debug.Assert(alignment > 0, "Alignment must be non-zero.");
+
+            long mask = alignment - 1;
+
+            Debug.Assert((alignment & mask) == 0, "Alignment should be a power of 2.");
+
+            return ((size + mask) & -mask);
+        }
+
+        public static long AlignSizeDown(long size, long alignment)
+        {
+            Debug.Assert(alignment > 0, "Alignment must be non-zero.");
+
+            long mask = alignment - 1;
+
+            Debug.Assert((alignment & mask) == 0, "Alignement should be a power of 2.");
+
+            return (size & -mask);
         }
     }
 }